Eksik süre ucunun DİZİ dalı canlıda tamamen bozuktu (Postgres 42804)
`POST /admin/eksik-sure` dizi için İSTİSNASIZ 500 veriyordu; film dalı
sağlamdı. Canlıda bağımsız doğrulandı: tv → 500, movie → 200 (aynı sahte id).

KÖK NEDEN: `SELECT DISTINCT 'tv', …, $2, 'elle'` içindeki TİPSİZ parametre.
node-pg parametreyi unknown gönderiyor; SELECT DISTINCT eşitlik operatörü
aramak için tipi HEDEF SÜTUNA bakmadan parse aşamasında text'e çözüyor, sonra
`dakika int` sütununa çarpıyor:
  column "dakika" is of type integer but expression is of type text
Film dalı VALUES kullandığı için etkilenmiyordu. Hata PARSE aşamasında, yani
0 satır eşleşse bile atıyor. Düzeltme: `$2::int`.

NEDEN TESTLER YAKALAMADI: eksik_sure.test.js SQL'i çalıştırmıyor, kaynağı
regex'liyor. Cast'in VARLIĞINI şart koşan regresyon testi eklendi (11/11).

SONUÇ (uç düzelince 80 dizinin süresi girildi, 5.240 izleme satırı):
  film %92,2 → %99,7 · dizi %72,9 → %97,9
  eksik yapım 450 → 21 · gerçek izleme süresi 2,93M → 4,42M dakika
Bunun büyük kısmı `sure_doldur.js`in önbellek + TMDB aşamalarından geldi;
elle giriş yalnız TMDB'nin de bilmediği yapımlar için kullanıldı.

Backend 1934/1934.

+++ backend/server.js
@@ -17573,8 +17573,20 @@ app.post('/admin/eksik-sure', adminKisit, sarici(async (req, res) => {
        VALUES ('movie', $1, 0, 0, $2, 'elle')
        ON CONFLICT (tur, tmdb_id, sezon, bolum) DO NOTHING`, [tmdbId, dakika])
     : await havuz.query(
+      // `$2::int` ŞART — çıplak `$2` bu ucu TAMAMEN çalışmaz yapıyordu (500,
+      // 27 Ağu 2026). Sebep: node-pg parametreyi TİPSİZ gönderir; `SELECT
+      // DISTINCT` eşitlik operatörü aramak için tipi HEDEF SÜTUNA bakmadan,
+      // parse aşamasında `unknown` -> `text` diye çözer. Sonra INSERT hedefi
+      // `dakika int` ile karşılaşınca Postgres 42804 verir:
+      //   column "dakika" is of type integer but expression is of type text
+      // Film dalı aynı hatayı YAŞAMAZ: orada `VALUES` var, hedef sütun tipi
+      // parametreye doğrudan uygulanır. Yani hata yalnız dizi dalındaydı ve
+      // 0 satır eşleşse bile atıyordu (parse hatası, çalıştırma değil).
+      // Testler yalnız kaynak metnini regex'liyor, SQL'i ÇALIŞTIRMIYOR —
+      // bu yüzden yakalanmadı; aşağıdaki cast'i kaldıran regresyon testi
+      // test/eksik_sure.test.js'e eklendi.
       `INSERT INTO yapim_sureleri (tur, tmdb_id, sezon, bolum, dakika, kaynak)
-       SELECT DISTINCT 'tv', i.tmdb_id, i.sezon, i.bolum, $2, 'elle'
+       SELECT DISTINCT 'tv', i.tmdb_id, i.sezon, i.bolum, $2::int, 'elle'
          FROM izlemeler i
          ${SURE_KAYNAK_JOIN}
         WHERE i.tur = 'tv' AND i.tmdb_id = $1 AND s.dakika IS NULL
